pyproject.toml
[project]
name = "murder-creek-barbell"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-allauth>=65.13.0",
    "django-cotton>=2.4.0",
    "django-crispy-forms>=2.5",
    "django-shinobi>=1.4.0",
    "stripe>=14.1.0",
]
